atoi

+++ ft_atoi.c
@@ -15,8 +15,8 @@
 int	ft_atoi(const char *str)
 {
 	int	i;
-	long long int	sign;
-	long long int	result;
+	int	sign;
+	int	result;
 
 	i = 0;
 	sign = 1;
